+++ Kernel/Storage/NVMe/NVMeController.cpp
@@ -11,7 +11,7 @@
 #include <AK/Types.h>
 #include <Kernel/Arch/SafeMem.h>
 #include <Kernel/Arch/x86/IO.h>
-#include <Kernel/Arch/x86/Processor.h>
+#include <Kernel/Arch/x86/x86Processor.h>
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Devices/Device.h>
